Remove temporary abstraction and old algorithm
