--- src/main/java/appeng/me/pathfinding/IPathingCalculation.java
@@ -1,10 +1,0 @@
-package appeng.me.pathfinding;
-
-public interface IPathingCalculation {
-    void step();
-
-    boolean isFinished();
-
-    int getChannelsByBlocks();
-    int getChannelsInUse();
-}
