+++ src/main/java/appeng/me/service/PathingService.java
@@ -21,9 +21,7 @@ package appeng.me.service;
 import java.util.HashSet;
 import java.util.Set;
 
-import appeng.me.pathfinding.CheckingPathingCalculation;
-import appeng.me.pathfinding.IPathingCalculation;
-import appeng.me.pathfinding.SmarterPathingCalculation;
+import appeng.me.pathfinding.PathingCalculation;
 import net.neoforged.neoforge.common.util.Lazy;
 import org.jetbrains.annotations.Nullable;
 
@@ -53,7 +51,6 @@ import appeng.me.Grid;
 import appeng.me.pathfinding.AdHocChannelUpdater;
 import appeng.me.pathfinding.ChannelFinalizer;
 import appeng.me.pathfinding.ControllerValidator;
-import appeng.me.pathfinding.PathingCalculation;
 
 public class PathingService implements IPathingService, IGridServiceProvider {
     private static final String TAG_CHANNEL_MODE = "cm";
@@ -66,7 +63,7 @@ public class PathingService implements IPathingService, IGridServiceProvider {
                 });
     }
 
-    private IPathingCalculation ongoingCalculation = null;
+    private PathingCalculation ongoingCalculation = null;
     private final Set<ControllerBlockEntity> controllers = new HashSet<>();
     private final Set<IGridNode> nodesNeedingChannels = new HashSet<>();
     private final Set<IGridNode> cannotCarryCompressedNodes = new HashSet<>();
@@ -130,10 +127,7 @@ public class PathingService implements IPathingService, IGridServiceProvider {
             } else if (this.controllerState == ControllerState.CONTROLLER_CONFLICT) {
                 this.grid.getPivot().beginVisit(new AdHocChannelUpdater(0));
             } else {
-                this.ongoingCalculation = new CheckingPathingCalculation(
-                        grid,
-                        Lazy.of(() -> new SmarterPathingCalculation(grid)),
-                        Lazy.of(() -> new PathingCalculation(grid)));
+                this.ongoingCalculation = new PathingCalculation(grid);
             }
         }
